deleted Controllers.java; moved content to ActivityController.java
